fix warnings

+++ src/binacpp_logger.cpp
@@ -29,7 +29,7 @@ BinaCPP_logger::write_log( const char *fmt, ... )
     struct tm * now = localtime( &t );
 
 
-    sprintf( new_fmt , "%04d-%02d-%02d %02d:%02d:%02d %06ld :%s\n" , now->tm_year + 1900, now->tm_mon + 1, now->tm_mday, now->tm_hour, now->tm_min, now->tm_sec , tv.tv_usec , fmt );
+    sprintf( new_fmt , "%04d-%02d-%02d %02d:%02d:%02d %06ld :%s\n" , now->tm_year + 1900, now->tm_mon + 1, now->tm_mday, now->tm_hour, now->tm_min, now->tm_sec , (long)tv.tv_usec , fmt );
 
     va_start (arg, fmt);
